Fix static checks deleting all .py files
In checking the notebooks, we turn them into .py files. We now
only delete the .py files associated with notebooks, and not
other .py files that might be in the `docs/examples` directory.
We also only run the `sed` command on those files, and test that
the correct files exist once the static script has completed.