Add pr-number script
When we make a changelog entry, we add a link to the pull request
for additional context. If the pull request doesn't exist yet,
we need to figure out what number the PR will end up being. This
script gets that number and formats it as a URL that can be used
in changelogs.

Co-authored-by: Eric Hunsberger <eric.hunsberger@appliedbrainresearch.com>