Add codespell_ignore_words option to static.sh
This is a list of words that codespell should ignore.
We ensure that it is a list using `check_list`.

Also removed the try/except around `check_list` because it
should not be possible to get a `KeyError` since we check
that the key is in the dictionary before doing any checks.
Additionally, the way it was written before was buggy because
it would stop checking later options if the KeyError was raised..