Factor out templating to separate file
In preparation for loading the templates to a string when
checking files by content rather than version string.

One important change is that section names and template filenames
are always automatically determined using the output filename,
which is the only information needed to get access to a template
now through the `BonesTemplate` class.

Co-authored-by: Daniel Rasmussen <daniel.rasmussen@appliedbrainresearch.com>