Error when file is missing from config
An error occurred in a downstream repo when they started from
a templated file, found that it was not sufficient for their
use case, removed the configuration from .nengobones.yml, but
kept the string 'Automatically generated by nengo-bones' in the
file. This commit detects that case, fails bones-check, and
instructs the user to either remove that text from the file
or add configuration to .nengobones.yml.

I found the bones-check tests to be a bit difficult to follow
due to it all being in one test with a bunch of state lying around,
so in order to write a test for this case I refactored that file
so that each case is tested in a separate test.

Addresses #61.