Add support for typed projects
This support is enabled with the `py_typed` section, which
does two things:

- Adds a {{ pkg_name }}/py.typed file, which lets other packages
  know that the project is typed.
- Adds mypy checking to static script and pre-commit hook.

The  py.typed file is the first templated file we've added that
should reside in the package directory. It required a little
bit of special case code; specifically, we treat the `templates/pkg`
directory as special when generating the section name (we don't
include the `pkg_` part) and when saving the file (we replace the
`pkg/` part of the path with `{{ pkg_name }}`).

Also:

- Added py.typed to MANIFEST.in and verified that when you
  create an sdist or bdist_wheel, the file shows up,
  as long as include_package_data==True
- My jupyter was adding a widgets metadata section, so I
  changed bones-format-notebook to remove it on formatting