Add all sections to CI script config
Previously CI scripts worked slightly differently from other
templates, in that they did not have access to the rest of the
.nengobones.yml file, only their own section, `pkg_name`, and
`repo_name`. This commit makes them more similar to other templates
by giving them access to the whole configuration file, with the
same promotion of that specific script's information.

This commit also adds two directories to the skip list for codespell.