Add main_branch option to main section
This option allows repos to use `main` (or anything else) as their
default git branch name, instead of the now discouraged `master`.
We still set the default value to `master` in order to minimize
the number of changes needed to downstream repos, but eventually
we should switch this to default to `main`.