Move license to top-level and add ABR license
The ABR license replaces the Nengo license. There are two
variants, abr-free and abr-nonfree. The license defaults to
"proprietary" now, which is likely what people want to be using
if they don't explicitly specify a license.

The license_rst section is still needed to render out the license
text, but the top-level license is also used in other templates.